pyproject.toml
[project]
name = "quant-bobe"
version = "0.1.0"
description = "Best-of-Breed Ensemble equities strategy with Alpaca paper trading"
requires-python = ">=3.11"
authors = [
  { name = "Quant BOBE" }
]
dependencies = [
  "pandas",
  "numpy",
  "scipy",
  "statsmodels",
  "pydantic",
  "pyyaml",
  "plotly",
  "matplotlib",
  "cvxpy",
  "loguru",
  "requests",
  "websockets",
  "python-dotenv",
  "yfinance",
  "alpaca-trade-api",
  "numba"
]

[project.optional-dependencies]
dev = ["black", "ruff", "mypy", "pytest"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "B", "I"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]
